Restore missing methods to InteractionController
